+++ core/class/AbeilleTools.class.php
@@ -760,9 +760,9 @@
                 $config['ab::zgIeeeAddrOk'.$zgId] = config::byKey('ab::zgIeeeAddrOk'.$zgId, 'Abeille', 0);
                 $config['ab::zgIeeeAddr'.$zgId] = config::byKey('ab::zgIeeeAddr'.$zgId, 'Abeille', '');
                 $chan = config::byKey('ab::zgChan'.$zgId, 'Abeille', 'niet');
-                $config['ab::zgPowerCycle'.$zgId] = config::byKey('ab::zgPowerCycle'.$zgId, 'Abeille', 1);
                 if ($chan != 'niet')
                     $config['ab::zgChan'.$zgId] = $chan;
+                $config['ab::preventUsbPowerCycle'] = config::byKey('ab::preventUsbPowerCycle', 'Abeille', 'N');
             }
             $config['ab::monitorId'] = config::byKey('ab::monitorId', 'Abeille', false);
             $config['ab::defaultParent'] = config::byKey('ab::defaultParent', 'Abeille', '1', 1);